refactor: relocate smooth pipeline to compressor/transform/smooth
Move SmoothQuant code into a new self-contained package with a shared
core/ layer; the vLLM online path and the HF offline convert path now
both consume the same algorithm primitives (formulas, QDQ, RoPE/GQA,
alpha searcher, smooth stats I/O) instead of carrying ~3000 lines of
duplicated implementation across vllm_calibrate_utils and
tools/smooth/convert_smooth_weights.py. Patched fused_moe.py now
imports from a standalone smooth_moe_inject module deployed by
install.sh next to vllm_calibrate_utils.
