+++ angelslim/compressor/quant/core/vllm_calibrate_utils/search.py
@@ -14,7 +14,7 @@ are kept module-private with the underscore prefix.
 """
 
 import os
-from dataclasses import dataclass
+from concurrent.futures import ThreadPoolExecutor, as_completed
 
 import torch
 
@@ -34,13 +34,6 @@ __all__ = [
     "remove_kvcache_perhead_value_hooks",
     "KVScaleSearcherPerHead",
     "get_kv_scale_search_results_perhead",
-    # Smooth alpha search
-    "SmoothAlphaSearchConfig",
-    "SmoothAlphaValueHook",
-    "SmoothAlphaSearcher",
-    "setup_smooth_alpha_search_hooks",
-    "remove_smooth_alpha_search_hooks",
-    "collect_fused_moe_alpha_search_values",
 ]
 
 
@@ -293,9 +286,6 @@ class KVScaleSearcher:
         self.num_steps = num_steps
 
     def __call__(self, model):
-        import os
-        from concurrent.futures import ThreadPoolExecutor, as_completed
-
         fp8_max = torch.finfo(torch.float8_e4m3fn).max  # 448.0
 
         # Collect raw kv tensors stored by the value hook
@@ -759,705 +749,3 @@ def get_kv_scale_search_results_perhead(results_list: list) -> dict:
         final[stats_key] = [head_dict[i] for i in sorted_indices]
 
     return final
-
-
-# =============================================================================
-# Smooth Alpha Grid Search (down_proj only)
-# =============================================================================
-
-
-@dataclass
-class SmoothAlphaSearchConfig:
-    """Configuration for smooth alpha grid search."""
-
-    alpha_min: float = 0.3
-    alpha_max: float = 1.0
-    alpha_steps: int = 8  # [0.3, 0.4, ..., 1.0]
-    act_quant_method: str = "per_token"  # "per_tensor" | "per_token"
-    act_quant_type: str = "int8"  # "int8" | "fp8"
-    weight_quant_method: str = (
-        "per_channel"  # "per_tensor" | "per_channel" | "per_group" | "per_block"
-    )
-    weight_quant_type: str = "int8"  # "int8" | "int4" | "fp8"
-    weight_quant_bits: int = 8
-    weight_group_size: int = 128  # per_group, -1 = per_channel
-    block_size: int = 128  # per_block fp8
-    use_ema_for_absmax: bool = False
-    smooth_search_mode: str = "default"  # "default" | "per-tensor-act-first"
-    act_mul_min: float = 0.1  # per-tensor-act-first: multiplier range min
-    act_mul_max: float = 1.0  # per-tensor-act-first: multiplier range max
-    smooth_min: float = 1e-6  # per-tensor-act-first: smooth clamp lower bound
-    smooth_max: float = 1e6  # per-tensor-act-first: smooth clamp upper bound
-
-
-# ---------------------------------------------------------------------------
-# QDQ helpers for grid search
-# ---------------------------------------------------------------------------
-
-
-def _smooth_qdq_act(
-    x: torch.Tensor,
-    method: str = "per_token",
-    quant_type: str = "int8",
-    bits: int = 8,
-) -> torch.Tensor:
-    """Quantize-dequantize activation (already divided by smooth_weight).
-
-    Args:
-        x: [N, in_features], float32
-        method: "per_tensor" | "per_token"
-        quant_type: "int8" | "fp8"
-        bits: bit width for INT
-    Returns:
-        x_qdq: [N, in_features], float32
-    """
-    if quant_type in ("int8", "int4", "int"):
-        bnt = (1 << (bits - 1)) - 1  # 127 for int8
-        if method == "per_tensor":
-            scale = x.abs().max().clamp(min=1e-8) / bnt
-        else:  # per_token
-            scale = x.abs().amax(dim=-1, keepdim=True).clamp(min=1e-8) / bnt
-        return (x / scale).round().clamp(-bnt - 1, bnt) * scale
-
-    elif quant_type == "fp8":
-        fp8_max = torch.finfo(torch.float8_e4m3fn).max  # 448.0
-        if method == "per_tensor":
-            scale = x.abs().max().clamp(min=1e-8) / fp8_max
-        else:  # per_token
-            scale = x.abs().amax(dim=-1, keepdim=True).clamp(min=1e-8) / fp8_max
-        q = (x / scale).clamp(-fp8_max, fp8_max).to(torch.float8_e4m3fn)
-        return q.float() * scale
-
-    else:
-        raise ValueError(f"Unknown act quant_type: {quant_type}")
-
-
-def _smooth_qdq_weight(
-    w: torch.Tensor,
-    method: str = "per_channel",
-    quant_type: str = "int8",
-    bits: int = 8,
-    group_size: int = 128,
-    block_size: int = 128,
-) -> torch.Tensor:
-    """Quantize-dequantize weight (already multiplied by smooth_weight).
-
-    Args:
-        w: [out_features, in_features], float32
-        method: "per_tensor" | "per_channel" | "per_group" | "per_block"
-        quant_type: "int8" | "int4" | "int" | "fp8"
-        bits: bit width for INT
-        group_size: for per_group (along dim=1)
-        block_size: for per_block fp8
-    Returns:
-        w_qdq: [out_features, in_features], float32
-    """
-    if quant_type in ("int8", "int4", "int"):
-        bnt = (1 << (bits - 1)) - 1
-
-        if method == "per_tensor":
-            scale = w.abs().max().clamp(min=1e-8) / bnt
-            return (w / scale).round().clamp(-bnt - 1, bnt) * scale
-
-        elif method == "per_channel":
-            # per output-channel (per row)
-            scale = w.abs().amax(dim=-1, keepdim=True).clamp(min=1e-8) / bnt
-            return (w / scale).round().clamp(-bnt - 1, bnt) * scale
-
-        elif method == "per_group":
-            # group along in_features (dim=1)
-            out_feat, in_feat = w.shape
-            gs = group_size if group_size > 0 else in_feat
-            if in_feat % gs != 0:
-                # pad to group_size multiple
-                pad_len = gs - in_feat % gs
-                w_padded = torch.nn.functional.pad(w, (0, pad_len))
-            else:
-                w_padded = w
-                pad_len = 0
-            w_grouped = w_padded.reshape(out_feat, -1, gs)
-            scale = w_grouped.abs().amax(dim=-1, keepdim=True).clamp(min=1e-8) / bnt
-            w_qdq = (w_grouped / scale).round().clamp(-bnt - 1, bnt) * scale
-            w_qdq = w_qdq.reshape(out_feat, -1)
-            if pad_len > 0:
-                w_qdq = w_qdq[:, :in_feat]
-            return w_qdq
-
-        else:
-            raise ValueError(f"Unknown weight method for INT: {method}")
-
-    elif quant_type == "fp8":
-        fp8_max = torch.finfo(torch.float8_e4m3fn).max
-
-        if method == "per_tensor":
-            scale = w.abs().max().clamp(min=1e-8) / fp8_max
-            q = (w / scale).clamp(-fp8_max, fp8_max).to(torch.float8_e4m3fn)
-            return q.float() * scale
-
-        elif method == "per_channel":
-            scale = w.abs().amax(dim=-1, keepdim=True).clamp(min=1e-8) / fp8_max
-            q = (w / scale).clamp(-fp8_max, fp8_max).to(torch.float8_e4m3fn)
-            return q.float() * scale
-
-        elif method == "per_block":
-            out_feat, in_feat = w.shape
-            # pad to block_size multiples
-            pad_r = (block_size - in_feat % block_size) % block_size
-            pad_b = (block_size - out_feat % block_size) % block_size
-            if pad_r > 0 or pad_b > 0:
-                w_padded = torch.nn.functional.pad(w, (0, pad_r, 0, pad_b))
-            else:
-                w_padded = w
-            po, pi = w_padded.shape
-            w_blocks = w_padded.reshape(po // block_size, block_size, pi // block_size, block_size)
-            w_blocks = w_blocks.permute(0, 2, 1, 3)  # [bo, bi, bs, bs]
-            scale = w_blocks.abs().amax(dim=(-2, -1), keepdim=True).clamp(min=1e-8) / fp8_max
-            q = (w_blocks / scale).clamp(-fp8_max, fp8_max).to(torch.float8_e4m3fn)
-            w_qdq = (q.float() * scale).permute(0, 2, 1, 3).reshape(po, pi)
-            return w_qdq[:out_feat, :in_feat]
-
-        elif method == "per_group":
-            # FP8 per_group: group along dim=1
-            out_feat, in_feat = w.shape
-            gs = group_size if group_size > 0 else in_feat
-            if in_feat % gs != 0:
-                pad_len = gs - in_feat % gs
-                w_padded = torch.nn.functional.pad(w, (0, pad_len))
-            else:
-                w_padded = w
-                pad_len = 0
-            w_grouped = w_padded.reshape(out_feat, -1, gs)
-            scale = w_grouped.abs().amax(dim=-1, keepdim=True).clamp(min=1e-8) / fp8_max
-            q = (w_grouped / scale).clamp(-fp8_max, fp8_max).to(torch.float8_e4m3fn)
-            w_qdq = (q.float() * scale).reshape(out_feat, -1)
-            if pad_len > 0:
-                w_qdq = w_qdq[:, :in_feat]
-            return w_qdq
-
-        else:
-            raise ValueError(f"Unknown weight method for FP8: {method}")
-
-    else:
-        raise ValueError(f"Unknown weight quant_type: {quant_type}")
-
-
-# ---------------------------------------------------------------------------
-# SmoothAlphaValueHook — capture raw activation for dense MLP down_proj
-# ---------------------------------------------------------------------------
-
-
-class SmoothAlphaValueHook:
-    """Capture raw down_proj input activation for alpha grid search.
-
-    Stores up to ``max_tokens`` token-rows per layer via CPU offload
-    with uniform sub-sampling when the cap is exceeded.
-
-    Works alongside SmoothDownProjInputHook (which collects absmax/ema
-    statistics); this hook stores the actual tensor values.
-
-    Args:
-        layer_name: down_proj module name (same key used in ``_smooth_stats``).
-        alpha_search_values: shared dict ``{layer_name: {"tokens": Tensor|None}}``.
-        max_tokens: maximum token-rows to keep per layer.
-    """
-
-    def __init__(self, layer_name, alpha_search_values, max_tokens=4096):
-        self.layer_name = layer_name
-        self.alpha_search_values = alpha_search_values
-        self.max_tokens = max_tokens
-
-    def __call__(self, module, input, output):
-        act = input[0] if isinstance(input, tuple) else input
-        if not isinstance(act, torch.Tensor):
-            return
-
-        with torch.no_grad():
-            act_cpu = act.detach().float().cpu()
-            stored = self.alpha_search_values[self.layer_name]
-            if stored["tokens"] is None:
-                stored["tokens"] = act_cpu
-            else:
-                stored["tokens"] = torch.cat([stored["tokens"], act_cpu], dim=0)
-            # Uniform sub-sample when cap exceeded
-            if stored["tokens"].shape[0] > self.max_tokens:
-                indices = torch.linspace(0, stored["tokens"].shape[0] - 1, self.max_tokens).long()
-                stored["tokens"] = stored["tokens"][indices]
-
-
-# ---------------------------------------------------------------------------
-# MoE raw tensor capture (kernel injection)
-# ---------------------------------------------------------------------------
-
-# Module-level cache for VLLM_MOE_COLLECT_ALPHA_SEARCH flag.
-# Initialised from env var so that *every* import of this module (including
-# the one inside fused_moe.py which is a separate module instance from the
-# angelslim.compressor.quant.core.vllm_calibrate_utils import) starts with
-# the correct value when the env var is set by the caller script.
-_MOE_COLLECT_ALPHA_SEARCH = os.getenv("VLLM_MOE_COLLECT_ALPHA_SEARCH", "0") == "1"
-
-
-def _set_moe_collect_alpha_search(enable: bool) -> None:
-    """Update the module-level MoE alpha-search collection flag."""
-    global _MOE_COLLECT_ALPHA_SEARCH
-    _MOE_COLLECT_ALPHA_SEARCH = bool(enable)
-
-
-def collect_fused_moe_alpha_search_values(
-    stage,
-    hidden_states,
-    topk_ids,
-    layer_name=None,
-    alpha_search_values=None,
-    max_tokens_per_expert=512,
-):
-    """Collect per-expert raw activation tensors for alpha grid search.
-
-    Called from inside the vLLM FusedMoE kernel (requires source modification),
-    at the same injection point as ``collect_fused_moe_smooth_stats``.
-
-    Args:
-        stage: must be "down_proj"
-        hidden_states: [num_tokens * top_k, intermediate_size_shard]
-        topk_ids: [num_tokens, top_k]
-        layer_name: FusedMoE layer name
-        alpha_search_values: model._alpha_search_values dict
-        max_tokens_per_expert: cap per expert
-    """
-
-    if alpha_search_values is None or layer_name is None:
-        return
-    if stage != "down_proj":
-        return
-
-    with torch.no_grad():
-        act = hidden_states
-        if act.dim() == 1:
-            act = act.unsqueeze(0)
-
-        num_tokens_hs = act.shape[0]
-        num_tokens_topk = topk_ids.shape[0]
-        top_k = topk_ids.shape[1]
-
-        if num_tokens_hs == num_tokens_topk * top_k:
-            flat_expert_ids = topk_ids.reshape(-1)
-        else:
-            return  # shape mismatch, skip
-
-        for expert_id_tensor in flat_expert_ids.unique():
-            expert_id = expert_id_tensor.item()
-            if expert_id < 0:
-                continue
-            expert_key = f"{layer_name}.{expert_id}.down_proj"
-            mask = flat_expert_ids == expert_id_tensor
-            expert_act = act[mask].detach().float().cpu()
-            if expert_act.numel() == 0:
-                continue
-
-            if expert_key not in alpha_search_values:
-                alpha_search_values[expert_key] = {"tokens": None}
-            stored = alpha_search_values[expert_key]
-            if stored["tokens"] is None:
-                stored["tokens"] = expert_act
-            else:
-                stored["tokens"] = torch.cat([stored["tokens"], expert_act], dim=0)
-            # Cap
-            if stored["tokens"].shape[0] > max_tokens_per_expert:
-                indices = torch.linspace(
-                    0, stored["tokens"].shape[0] - 1, max_tokens_per_expert
-                ).long()
-                stored["tokens"] = stored["tokens"][indices]
-
-
-# ---------------------------------------------------------------------------
-# setup / remove alpha search hooks
-# ---------------------------------------------------------------------------
-
-
-def setup_smooth_alpha_search_hooks(
-    model,
-    max_tokens=4096,
-    collect_moe=True,
-):
-    """Register hooks to capture raw down_proj activations for alpha grid search.
-
-    Must be called via ``llm.apply_model()`` AFTER ``setup_smooth_hooks()``.
-
-    Dense MLP: registers SmoothAlphaValueHook on each down_proj layer.
-    MoE: wires w13_weight attributes so that ``collect_fused_moe_alpha_search_values``
-         (called from the kernel injection point) stores raw tensors per expert.
-
-    Returns a summary string.
-    """
-    from vllm.model_executor.layers.linear import LinearBase
-
-    if not hasattr(model, "_alpha_search_values"):
-        model._alpha_search_values = {}
-    if not hasattr(model, "_alpha_search_hooks"):
-        model._alpha_search_hooks = []
-
-    # --- Dense MLP down_proj ---
-    all_linear = _find_layers(model, layers=[torch.nn.Linear, LinearBase])
-    down_proj_layers = {
-        name: module for name, module in all_linear.items() if name.split(".")[-1] == "down_proj"
-    }
-
-    for name, layer in down_proj_layers.items():
-
-        if name not in model._alpha_search_values:
-            model._alpha_search_values[name] = {"tokens": None}
-        hook = SmoothAlphaValueHook(name, model._alpha_search_values, max_tokens)
-        handle = layer.register_forward_hook(hook)
-        model._alpha_search_hooks.append(handle)
-
-    # --- MoE (kernel injection) ---
-    moe_count = 0
-    if collect_moe:
-        try:
-            from vllm.model_executor.layers.fused_moe.layer import FusedMoE
-
-            moe_layers = _find_layers(model, layers=[FusedMoE])
-        except ImportError:
-            moe_layers = {}
-
-        if moe_layers:
-            os.environ["VLLM_MOE_COLLECT_ALPHA_SEARCH"] = "1"
-            _set_moe_collect_alpha_search(True)
-
-        per_expert_max = max(1, max_tokens // 8)  # for topk=8
-        for name, layer in moe_layers.items():
-
-            if hasattr(layer, "w13_weight") and layer.w13_weight is not None:
-                layer.w13_weight._alpha_search_values_of_model = model._alpha_search_values
-                layer.w13_weight._vllm_layer_name_smooth = name
-                layer.w13_weight._alpha_search_max_tokens = per_expert_max
-                moe_count += 1
-            else:
-                print(
-                    f"[DEBUG] Cannot set smooth attrs on w13_weight for {name}: "
-                    f"hasattr={hasattr(layer, 'w13_weight')}, "
-                    f"is_none={getattr(layer, 'w13_weight', None) is None}"
-                )
-    total = len(model._alpha_search_hooks)
-    msg = (
-        f"Registered {total} alpha-search hooks (dense down_proj), "
-        f"{moe_count} MoE layers wired"
-    )
-    print(f"[Alpha Search] {msg}")
-    return msg
-
-
-def remove_smooth_alpha_search_hooks(model):
-    """Remove alpha-search hooks and free stored activation tensors."""
-    if hasattr(model, "_alpha_search_hooks"):
-        for handle in model._alpha_search_hooks:
-            handle.remove()
-        model._alpha_search_hooks.clear()
-    if hasattr(model, "_alpha_search_values"):
-        del model._alpha_search_values
-    os.environ.pop("VLLM_MOE_COLLECT_ALPHA_SEARCH", None)
-    _set_moe_collect_alpha_search(False)
-    return "Alpha-search hooks removed"
-
-
-# ---------------------------------------------------------------------------
-# _get_down_proj_weight — resolve weight tensor for dense / MoE
-# ---------------------------------------------------------------------------
-
-
-def _get_down_proj_weight(model, layer_key):
-    """Get the down_proj weight tensor (TP-local shard) for a given layer key.
-
-    Dense MLP key: "model.layers.0.mlp.down_proj"
-      → model.get_submodule(key).weight   [out, in_shard]
-
-    MoE expert key: "model.layers.3.mlp.experts.42.down_proj"
-      → parse moe_layer = "model.layers.3.mlp.experts"
-      → expert_id = 42
-      → FusedMoE.w2_weight[expert_id]     [out, in_shard]
-
-    Returns weight tensor (detached) or None.
-    """
-    import re
-
-    # Try direct submodule access (works for dense MLP)
-    try:
-        mod = model.get_submodule(layer_key)
-        if hasattr(mod, "weight"):
-            return mod.weight.detach()
-    except (AttributeError, ValueError):
-        pass
-
-    # MoE pattern: *.experts.<expert_id>.down_proj
-    moe_match = re.match(r"^(.+)\.(\d+)\.down_proj$", layer_key)
-    if moe_match:
-        moe_prefix = moe_match.group(1)  # e.g. "model.layers.3.mlp.experts"
-        expert_id = int(moe_match.group(2))
-        try:
-            moe_module = model.get_submodule(moe_prefix)
-        except (AttributeError, ValueError):
-            # Try one level up (experts might be inside FusedMoE)
-            parent = moe_prefix.rsplit(".", 1)[0] if "." in moe_prefix else moe_prefix
-            try:
-                moe_module = model.get_submodule(parent)
-            except (AttributeError, ValueError):
-                return None
-
-        # FusedMoE stores w2 (down_proj) as w2_weight: [num_experts, out, in_shard]
-        w2 = getattr(moe_module, "w2_weight", None)
-        if w2 is not None and expert_id < w2.shape[0]:
-            return w2[expert_id].detach()
-
-    return None
-
-
-# ---------------------------------------------------------------------------
-# SmoothAlphaSearcher — core grid search callable
-# ---------------------------------------------------------------------------
-
-
-class SmoothAlphaSearcher:
-    """Per-layer alpha grid search for SmoothQuant down_proj.
-
-    Must be called via ``llm.apply_model()`` **before** ``get_smooth_stats()``
-    so that ``model._smooth_stats`` still contains the TP-local shard of absmax.
-
-    Each rank independently searches its own shard, then ``all_gather`` is
-    used to concatenate ``smooth_weight`` shards into the full
-    ``[intermediate_size]`` vector.
-
-    Usage::
-
-        searcher = SmoothAlphaSearcher(config, output_path="path/to/output.json")
-        summary_list = llm.apply_model(searcher)
-        # Results are saved to output_path by rank 0 inside the worker.
-    """
-
-    def __init__(self, config: SmoothAlphaSearchConfig, output_path: str = None):
-        self.config = config
-        self.output_path = output_path
-
-    def __call__(self, model):
-        import torch.distributed as dist
-
-        rank, world_size = _get_dist_info()
-        smooth_stats = getattr(model, "_smooth_stats", {})
-        alpha_values = getattr(model, "_alpha_search_values", {})
-
-        if not alpha_values:
-            print("[SmoothAlphaSearcher] No activation tensors collected, skipping.")
-            return {}
-
-        cfg = self.config
-        alphas = torch.linspace(cfg.alpha_min, cfg.alpha_max, cfg.alpha_steps).tolist()
-        use_gpu = torch.cuda.is_available()
-        device = (
-            torch.device("cuda", torch.cuda.current_device()) if use_gpu else torch.device("cpu")
-        )
-
-        results = {}
-        processed = 0
-        skipped = 0
-
-        for layer_key, stored in alpha_values.items():
-            x_cpu = stored.get("tokens")
-            if x_cpu is None or x_cpu.shape[0] == 0:
-                skipped += 1
-                continue
-
-            # --- Get weight ---
-            print(f"[SmoothAlphaSearcher] Processing layer {layer_key}")
-            w_tensor = _get_down_proj_weight(model, layer_key)
-            if w_tensor is None:
-                skipped += 1
-                continue
-
-            # --- Get local shard of absmax (pre-all_gather) ---
-            stat_key = layer_key
-            stat = smooth_stats.get(stat_key)
-            if stat is None or stat.get("absmax") is None:
-                skipped += 1
-                continue
-            local_absmax = stat["ema" if cfg.use_ema_for_absmax else "absmax"]
-            if local_absmax is None:
-                local_absmax = stat["absmax"]
-            if local_absmax is None:
-                skipped += 1
-                continue
-
-            # --- Move to device ---
-            x = x_cpu.to(device)
-            w = w_tensor.float().to(device)
-            act_absmax = local_absmax.float().to(device).clamp(min=1e-8)
-            weight_absmax = w.abs().max(dim=0).values.clamp(min=1e-8)
-
-            # --- Reference output (computed once) ---
-            y_ref = x @ w.T  # [N, out_feat]
-
-            # --- Grid search no tensor-wise activation mode ---
-
-            losses = torch.empty(len(alphas), device=device)
-            if cfg.smooth_search_mode == "default":
-                for i, alpha in enumerate(alphas):
-                    smooth = (act_absmax.pow(alpha) / weight_absmax.pow(1.0 - alpha)).clamp(
-                        1e-6, 1e6
-                    )
-
-                    x_smoothed = x / smooth.unsqueeze(0)
-                    w_smoothed = w * smooth.unsqueeze(0)
-
-                    x_q = _smooth_qdq_act(
-                        x_smoothed,
-                        method=cfg.act_quant_method,
-                        quant_type=cfg.act_quant_type,
-                        bits=cfg.weight_quant_bits,
-                    )
-                    w_q = _smooth_qdq_weight(
-                        w_smoothed,
-                        method=cfg.weight_quant_method,
-                        quant_type=cfg.weight_quant_type,
-                        bits=cfg.weight_quant_bits,
-                        group_size=cfg.weight_group_size,
-                        block_size=cfg.block_size,
-                    )
-                    y_q = x_q @ w_q.T
-                    losses[i] = ((y_ref - y_q) ** 2).mean()
-
-            # --- Grid search tensor-wise activation mode ---
-            # this mode search multipy with activation abs_max(per-tensor)
-            elif cfg.smooth_search_mode == "per-tensor-act-first":
-                per_tensor_act_absmax = act_absmax.max()
-                mul_list = torch.linspace(
-                    cfg.act_mul_min, cfg.act_mul_max, cfg.alpha_steps
-                ).tolist()
-                for i, mul in enumerate(mul_list):
-                    tgt_absmax = mul * per_tensor_act_absmax
-                    # calculate smooth weight
-                    # smooth all act channel to tgt_absmax
-                    smooth = act_absmax / tgt_absmax
-
-                    # not too big for smooth
-                    smooth = smooth.clamp(cfg.smooth_min, cfg.smooth_max)
-                    x_smoothed = x / smooth.unsqueeze(0)
-                    w_smoothed = w * smooth.unsqueeze(0)
-
-                    x_q = _smooth_qdq_act(
-                        x_smoothed,
-                        method=cfg.act_quant_method,
-                        quant_type=cfg.act_quant_type,
-                        bits=cfg.weight_quant_bits,
-                    )
-                    w_q = _smooth_qdq_weight(
-                        w_smoothed,
-                        method=cfg.weight_quant_method,
-                        quant_type=cfg.weight_quant_type,
-                        bits=cfg.weight_quant_bits,
-                        group_size=cfg.weight_group_size,
-                        block_size=cfg.block_size,
-                    )
-                    y_q = x_q @ w_q.T
-                    losses[i] = ((y_ref - y_q) ** 2).mean()
-
-            best_idx = int(losses.argmin().item())
-
-            if cfg.smooth_search_mode == "per-tensor-act-first":
-                best_mul = mul_list[best_idx]
-                best_tgt = best_mul * per_tensor_act_absmax
-                best_smooth = (act_absmax / best_tgt).clamp(cfg.smooth_min, cfg.smooth_max)
-                best_alpha = best_mul  # store the mul factor as "alpha" for this mode
-            else:
-                best_alpha = alphas[best_idx]
-                best_smooth = (
-                    act_absmax.pow(best_alpha) / weight_absmax.pow(1.0 - best_alpha)
-                ).clamp(cfg.smooth_min, cfg.smooth_max)
-
-            results[layer_key] = {
-                "alpha": best_alpha,
-                "smooth_weight_shard": best_smooth.cpu(),
-                "loss": losses[best_idx].item(),
-            }
-            processed += 1
-
-            # Free GPU memory (keep tensors on CPU, but don't call
-            # empty_cache inside the loop — it's very slow and can
-            # interfere with vLLM's memory allocator).
-            del x, w, y_ref, act_absmax, weight_absmax, x_cpu
-
-        print(
-            f"[SmoothAlphaSearcher] rank={rank}: processed={processed}, "
-            f"skipped={skipped}, total_keys={len(alpha_values)}"
-        )
-
-        # --- all_gather: concatenate shards across TP ranks ---
-        if world_size > 1 and dist.is_initialized():
-            # Batch all_gather: stack all smooth_weight shards into one tensor,
-            # do a single all_gather, then split back. This avoids thousands of
-            # individual NCCL calls which can trigger cudaErrorIllegalAddress.
-            keys_list = sorted(results.keys())
-            if keys_list:
-                # --- smooth_weight shards ---
-                sw_list = [results[k]["smooth_weight_shard"] for k in keys_list]
-                sw_sizes = [s.numel() for s in sw_list]
-                sw_flat = torch.cat([s.reshape(-1) for s in sw_list]).cuda()
-                gathered_flat = [torch.zeros_like(sw_flat) for _ in range(world_size)]
-                dist.all_gather(gathered_flat, sw_flat)
-                # concat along TP dim for each key
-                offset = 0
-                for i, key in enumerate(keys_list):
-                    sz = sw_sizes[i]
-                    chunks = [g[offset : offset + sz].cpu() for g in gathered_flat]
-                    results[key]["smooth_weight"] = torch.cat(chunks, dim=0).tolist()
-                    offset += sz
-
-                # --- alpha values ---
-                alpha_flat = torch.tensor(
-                    [results[k]["alpha"] for k in keys_list],
-                    dtype=torch.float32,
-                ).cuda()
-                gathered_alpha = [torch.zeros_like(alpha_flat) for _ in range(world_size)]
-                dist.all_gather(gathered_alpha, alpha_flat)
-                for i, key in enumerate(keys_list):
-                    results[key]["alpha"] = [g[i].item() for g in gathered_alpha]
-
-                for key in keys_list:
-                    results[key].pop("smooth_weight_shard", None)
-
-                del sw_flat, gathered_flat, alpha_flat, gathered_alpha
-                if use_gpu:
-                    torch.cuda.empty_cache()
-
-            dist.barrier()
-        else:
-            for key in results:
-                entry = results[key]
-                entry["smooth_weight"] = entry.pop("smooth_weight_shard").tolist()
-                entry["alpha"] = [entry["alpha"]]
-
-        # --- Save results to file on rank 0 (avoids pickling the huge
-        # dict back through the mp executor which can OOM / hang). ---
-        num_results = len(results)
-        if self.output_path and rank == 0 and results:
-            import json as _json
-
-            output_data = {
-                "config": {
-                    "alpha_min": cfg.alpha_min,
-                    "alpha_max": cfg.alpha_max,
-                    "alpha_steps": cfg.alpha_steps,
-                    "act_quant_method": cfg.act_quant_method,
-                    "act_quant_type": cfg.act_quant_type,
-                    "weight_quant_method": cfg.weight_quant_method,
-                    "weight_quant_type": cfg.weight_quant_type,
-                    "weight_quant_bits": cfg.weight_quant_bits,
-                },
-                "results": results,
-            }
-            os.makedirs(os.path.dirname(self.output_path) or ".", exist_ok=True)
-            with open(self.output_path, "w") as f:
-                _json.dump(output_data, f, indent=2)
-            print(
-                f"[SmoothAlphaSearcher] rank=0 saved {num_results} results to {self.output_path}"
-            )
-
-        # Return only a lightweight summary (NOT the full results dict)
-        # to avoid serialisation issues with mp executor.
-        return f"processed={num_results}"
